add rules

+++ src/main/java/com/abuse/module/Evaluation.java
@@ -3,22 +3,22 @@ package com.abuse.module;
 public enum Evaluation {
     EQUALS {
         @Override
-        boolean match(Long source, Long target) {
-            return source.longValue() == target.longValue();
+        boolean match(long source, Long target) {
+            return source == target.longValue();
         }
     }
     , LOWER_THAN_AND_EQUALS {
         @Override
-        boolean match(Long source, Long target) {
-            return source.longValue() >= target.longValue();
+        boolean match(long source, Long target) {
+            return source >= target.longValue();
         }
     }
     , GREATER_THAN_AND_EQUALS {
         @Override
-        boolean match(Long source, Long target) {
-            return source.longValue() <= target.longValue();
+        boolean match(long source, Long target) {
+            return source <= target.longValue();
         }
     }
     ;
-    abstract boolean match(Long source, Long target);
+    abstract boolean match(long source, Long target);
 }
